Запись результатов обработки PNG в PNG и сжатие

+++ retry.js
@@ -53,16 +53,23 @@ if (err) throw err;
 	}
 	//find_plain(image, 8, 20);
 	var extData = /([^\.]+)\.([^\.]+)$/.exec(img.originalname).slice(1,3);//[name,extension]
-	image.write('./uploads/'+extData[0] + '_mod.bmp', function() {
-		if (extData[1].toLowerCase() == 'jpg' || extData[1].toLowerCase() == 'jpeg') {
-			childProcess.execSync('convert ' +'./uploads/'+ extData[0] + '_mod.bmp -quality ' +
-			childProcess.execSync('identify -format \'%Q\' ' + img.path).toString() + ' ' +'./uploads/'+ extData[0] + '_mod.jpg');
-			//console.log(Date.now()-t);
-		}
-		else {
-			childProcess.execSync('convert ' +'./uploads/'+ extData[0] + '_mod.bmp ' +'./uploads/'+ extData[0] + '_mod.jpg');
-			//console.log(Date.now()-t);
-		}
-		childProcess.execSync('rm ' +'./uploads/'+ extData[0] + '_mod.bmp');
-	});
+	if (extData[1].toLowerCase() == 'png'){
+		image.write('./uploads/'+extData[0] + '_mod.png', function() {
+			childProcess.execSync('convert -depth 24 -define png:compression-level=9 ' +'./uploads/'+ extData[0] + '_mod.png ' +'./uploads/'+ extData[0] + '_mod.png');
+		});
+	}
+	else {
+		image.write('./uploads/'+extData[0] + '_mod.bmp', function() {
+			if (extData[1].toLowerCase() == 'jpg' || extData[1].toLowerCase() == 'jpeg') {
+				childProcess.execSync('convert ' +'./uploads/'+ extData[0] + '_mod.bmp -quality ' +
+				childProcess.execSync('identify -format \'%Q\' ' + img.path).toString() + ' ' +'./uploads/'+ extData[0] + '_mod.jpg');
+				//console.log(Date.now()-t);
+			}
+			else {
+				childProcess.execSync('convert ' +'./uploads/'+ extData[0] + '_mod.bmp ' +'./uploads/'+ extData[0] + '_mod.jpg');
+				//console.log(Date.now()-t);
+			}
+			childProcess.execSync('rm ' +'./uploads/'+ extData[0] + '_mod.bmp');
+		});
+	}
 });
